performance updates

+++ snapshot.py
@@ -35,13 +35,14 @@ def snapshot(directory, current_house):
 
     hash_digest = snapshot_hash.hexdigest()
     snapshot_data['file_list'] = list(snapshot_data['files'].keys())
-    update_version(snapshot_directory)
-    save_at = f'{snapshot_directory}/{current_version(snapshot_directory)}'
-    os.makedirs(save_at)
-    with open(f'{save_at}/{hash_digest}', 'wb') as f:
-      pickle.dump(snapshot_data, f)
-
-    print(f'Snapshot created with hash {hash_digest}')
+    if not os.path.exists(f'{snapshot_directory}/{current_version(snapshot_directory)}/{hash_digest}'):
+      save_at = f'{snapshot_directory}/{update_version(snapshot_directory)}'
+      os.makedirs(save_at)
+      with open(f'{save_at}/{hash_digest}', 'wb') as f:
+        pickle.dump(snapshot_data, f)
+      print(f'Snapshot created with hash {hash_digest}')
+    else:
+      print(f'Snapshot already exists')
     empty_ready_folder(ready_directory)
   else:
     print("Nothing to Snapshot")
@@ -53,8 +54,9 @@ def current_version(snapshot_directory):
   return version
 
 def update_version(snapshot_directory):
-  val = current_version(snapshot_directory)
-  val = int(val) + 1
+  version = current_version(snapshot_directory)
+  version = int(version) + 1
   version_file = open(f'{snapshot_directory}/version.txt', 'w')
-  version_file.write(f'{val}')
+  version_file.write(f'{version}')
   version_file.close()
+  return version
